Fix #200 where swagger used the wrong IP to retrieve well-known info
Swagger auth was using the internal docker IP. This won't work as you need to reference localhost:5005. 
Identity and API also weren't running on 5001/5000 and 5004/5005 as aspnetcore_urls were not referencing the ports correctly. 
In addition to that, I removed the part where direct ip's were referenced instead of service names. This change makes things easier to follow and less likely to break in the future, let Docker do it's DNS magic to resolve the IP address from the container.