Likes now have a datetime and retrieving projects returns likes
Likes were not being returned when retrieving projects. Due to performance reasons, we are just returning an array for now with user id + date instead of Like Count and boolean to check if currently logged in user has liked that specific project. 
In the future we should probably move the like and follow  things into their own controller with respective services. The frontend can then make a call to retrieve project meta data, which can take a little longer and should be non-blocking for the frontend. 
For now, the frontend can count the objects in the likes array and check if a user id is equal to the one that is logged in.
Fixes #329.